implement AuthorRepository with Spring Data Jpa
TODO: narrow service layer tests in TestServiceConfiguration
