+++ library/README.md
@@ -23,3 +23,8 @@ Homework for https://otus.ru/ Spring Developer course
 Добавить комментарии к книгам, и высокоуровневые сервисы, оставляющие комментарии к книгам.
 
 Покрыть DAO тестами используя H2 базу данных и соответствующий H2 Hibernate-диалект
+
+## HW7. "Белая магия" Spring Data: Spring Data JPA 
+
+Билиотеку на Spring Data JPA  
+Реализовать весь функционал работы с БД в приложении книг с использованием spring-data-jpa репозиториев.
