+++ library/src/test/java/cyclone/otusspring/library/service/TestServiceConfiguration.java
@@ -1,14 +1,19 @@
 package cyclone.otusspring.library.service;
 
-import org.springframework.boot.SpringBootConfiguration;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.context.annotation.ComponentScan;
+//import org.springframework.boot.SpringBootConfiguration;
+//import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+//import org.springframework.boot.autoconfigure.domain.EntityScan;
+//import org.springframework.context.annotation.ComponentScan;
+//import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
-@SpringBootConfiguration
-@EnableAutoConfiguration
-//@EnableConfigurationProperties (if needed)
-@ComponentScan({"cyclone.otusspring.library.repository.jpa", "cyclone.otusspring.library.service"})
-@EntityScan({"cyclone.otusspring.library.model"})
+//TODO narrow service tests context
+//@SpringBootConfiguration
+//@EnableAutoConfiguration
+//@EnableJpaRepositories
+////@EnableConfigurationProperties (if needed)
+//@ComponentScan({"cyclone.otusspring.library.repository"
+//        , "cyclone.otusspring.library.repository.datajpa"
+//        , "cyclone.otusspring.library.service"})
+//@EntityScan({"cyclone.otusspring.library.model"})
 public class TestServiceConfiguration {
 }
